requirements.txt
pynput==1.7.6
python-vlc==3.0.16120
selenium==4.2.0
